requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
alembic>=1.13.1
boto3>=1.34.14
opencv-python>=4.9.0
scikit-image>=0.21.0
scipy==1.15.2
numpy>=1.26.0
pillow>=10.2.0
email-validator>=2.1.0
requests>=2.31.0